pyproject.toml
[project]
name = "web-scrapping-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4",
    "request",
    "streamlit",
]
